requirements.txt
streamlit==1.27.0
python-docx==0.8.11
PyPDF2==3.0.1 